+++ src/app.tsx
@@ -1,6 +1,6 @@
 import "./i18n";
 import { Certificates } from "./components/certificates";
-import { Providers } from "./components/providers/Providers";
+import { Providers } from "./components/providers";
 
 export function App() {
   return (
