+++ src/components/approve-connection/index.ts
@@ -1,0 +1,1 @@
+export * from "./ApproveConnection";
