+++ src/components/certificates-providers-list-item/index.ts
@@ -1,0 +1,1 @@
+export * from "./CertificatesProvidersListItem";
