+++ src/components/certificates-providers-list-item/styles/index.module.scss
@@ -1,3 +1,3 @@
-.currentListItem {
+.current_list_item {
   color: var(--pv-color-primary);
 }
