+++ src/components/error-connection/ErrorConnection.tsx
