+++ src/components/error-connection/index.ts
@@ -1,0 +1,1 @@
+export * from "./ErrorConnection";
